feat: Reorder output of `semgrep ci` to allow the app to mark specific findings as blocking. (semgrep/semgrep-proprietary#2782)
**Important Note**
This PR will make the `semgrep ci` command take longer to output finding
data. Notably, *the full scan time remains unchanged*. It's just that
the point during the scan at which we print findings will now be later.

Right now whether or not a CI scan should exit with code 1 is determined
mostly by the CLI, but the app can use the `app_block_override` field of
`ci_scan_complete_respons`e to tell the CLI to exit with code 1 no
matter what. This is currently used to support blocking scans when we
detect a dependency with a blocked license.

We're currently working on a project which includes making it more
customizable within the app whether or not a given finding should be
blocking, right now just for supply chain, but with the aim to
eventually move the computation of blocking findings entirely to the
app. So we'd like to be able to tell the CLI not only "please block" but
also "here are the list of findings that the app determined are
blocking".

In order to support this, we need to re-order some of the CLI output.
Here's the steps involved in a CI scan right now:
1. CLI sends scan request to the app
2. App sends a scan config to the CLI. Some of the rules may be
annotated with `"dev.semgrep.actions": ["block"]` in their metadata
field
3. CLI runs the scan, gets findings. 
4. CLI uses the rule metadata described previously to determine which
findings are blocking
5. CLI outputs findings, *grouped by product and blocking status*
6. CLI sends findings (and other stuff, like dependencies) to the app
8. App starts processing findings and dependencies
9. App sends back "ok I got the findings"
10. CLI starts polling the app with "did you finish processing yet?"
11. Once it's done, the app sends back "ok I finished processing", and
optionally a field called `app_block_override`, which forces the CLI to
exit with code 1 if it's `true`. Right now this is *only* used to make
the CLI block if we scanned a dependency whose software license violated
the SCA license policy that the user has set up in the app
12. CLI exits

Here's how it will work after this PR, with changed things in **bold**
1. CLI sends scan request to the app
2. App sends a scan config to the CLI. Some of the rules may be
annotated with `"dev.semgrep.actions": ["block"]` in their metadata
field
3. CLI runs the scan, gets findings. 
4. CLI uses the rule metadata described previously to determine which
findings are blocking **(according to the old rules for what counts as
blocking, which are still supported)**
5. **NO FINDINGS ARE OUTPUT TO THE TERMINAL**
CLI sends findings (and other stuff, like dependencies) to the app
8. App starts processing findings and dependencies
9. App sends back "ok I got the findings"
10. CLI starts polling the app with "did you finish processing yet?"
11. Once it's done, the app sends back "ok I finished processing", and
optionally a field called `app_block_override`, which forces the CLI to
exit with code 1 if it's `true`. **It also optionally sends back a field
called `app_blocking_match_based_ids`, which are a list of findings that
the app determined should be blocking**
12. **The set of blocking match based ids from the app are used to
update the blocking status of all the findings from semgrep**
13. CLI outputs findings, *grouped by product and blocking status*,
**where blocking status now includes the information about which
findings the app thinks are blocking**
15. CLI exits

synced from Pro 196bd2343b70068b9b596473ba7bcc719eb668f1